Update rsync command (#33)

+++ {{cookiecutter.plugin_name}}/move_template_files.sh
@@ -1,4 +1,9 @@
 #!/bin/sh
 
-rsync -avh {{cookiecutter.plugin_name}}/ .
-rm -rfv {{cookiecutter.plugin_name}}
+if ! command -v rsync &> /dev/null; then
+  echo "rsync required, but not installed!"
+  exit 1
+else
+  rsync -avh {{cookiecutter.plugin_name}}/ .
+  rm -rfv {{cookiecutter.plugin_name}}
+fi
